+++ lib/kfx_dependency_injection/i_must_be.dart
@@ -1,0 +1,5 @@
+part of kfx_dependency_injection;
+
+abstract class IMustBeTransient {}
+
+abstract class IMustBeSingleton {}