+++ test/kfx_dependency_injection_test.dart
@@ -16,6 +16,14 @@ class _MockTestClass extends _TestClass {
   }
 }
 
+abstract class IRequireToBeSingleton implements IMustBeSingleton {}
+
+abstract class IRequireToBeTransient implements IMustBeTransient {}
+
+class MeSingleton implements IRequireToBeSingleton {}
+
+class MeTransient implements IRequireToBeTransient {}
+
 void _expectException<TException extends Exception>(void Function() method, String expectedMessage) {
   try {
     method();
@@ -30,111 +38,132 @@ void main() {
   tearDown(() => ServiceProvider.unregisterAll());
 
   test("Registered transient services must be transient", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "1");
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "2");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "1");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "2");
 
-    final tc1 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "1");
-    final tc2 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "2");
+    final tc1 = ServiceProvider.required<_TestClass>(key: "1");
+    final tc2 = ServiceProvider.required<_TestClass>(key: "2");
 
     tc1.value = 1;
     tc2.value = 2;
 
     expect(tc1.value, 1);
     expect(tc2.value, 2);
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "1").value, 0);
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "2").value, 0);
+    expect(ServiceProvider.required<_TestClass>(key: "1").value, 0);
+    expect(ServiceProvider.required<_TestClass>(key: "2").value, 0);
   });
 
   test("Overrides must work", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "1");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "1");
 
-    ServiceProvider.replace<_TestClass>((serviceProvider, platformInfo) => _MockTestClass(value: 11), key: "1");
-    ServiceProvider.replace<_TestClass>((serviceProvider, platformInfo) => _MockTestClass(value: 22), key: "2");
+    ServiceProvider.replace<_TestClass>((optional, required, platform) => _MockTestClass(value: 11), key: "1");
+    ServiceProvider.replace<_TestClass>((optional, required, platform) => _MockTestClass(value: 22), key: "2");
 
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "2");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "2");
 
-    final tc1 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "1");
-    final tc2 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "2");
+    final tc1 = ServiceProvider.required<_TestClass>(key: "1");
+    final tc2 = ServiceProvider.required<_TestClass>(key: "2");
 
     expect(tc1.value, 11);
     expect(tc2.value, 22);
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "1").value, 11);
+    expect(ServiceProvider.required<_TestClass>(key: "1").value, 11);
 
     tc1.value = 33;
 
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "2").value, 22);
+    expect(ServiceProvider.required<_TestClass>(key: "2").value, 22);
   });
 
   test("Registered singleton services must be singleton", () {
-    ServiceProvider.registerSingleton((serviceProvider, platformInfo) => _TestClass(), key: "1");
-    ServiceProvider.registerSingleton((serviceProvider, platformInfo) => _TestClass(), key: "2");
+    ServiceProvider.registerSingleton((optional, required, platform) => _TestClass(), key: "1");
+    ServiceProvider.registerSingleton((optional, required, platform) => _TestClass(), key: "2");
 
-    final tc1 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "1");
-    final tc2 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "2");
+    final tc1 = ServiceProvider.required<_TestClass>(key: "1");
+    final tc2 = ServiceProvider.required<_TestClass>(key: "2");
 
     tc1.value = 1;
     tc2.value = 2;
 
     expect(tc1.value, 1);
     expect(tc2.value, 2);
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "1").value, 1);
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "2").value, 2);
+    expect(ServiceProvider.required<_TestClass>(key: "1").value, 1);
+    expect(ServiceProvider.required<_TestClass>(key: "2").value, 2);
     tc2.value = 4;
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>(key: "2").value, 4);
+    expect(ServiceProvider.required<_TestClass>(key: "2").value, 4);
+  });
+
+  test("IMustBeTransient and IMustBeSingleton must validate the service", () {
+    ServiceProvider.registerTransient<IRequireToBeSingleton>((optional, required, platform) => MeSingleton());
+    ServiceProvider.registerSingleton<IRequireToBeTransient>((optional, required, platform) => MeTransient());
+
+    _expectException<RegistryMustBeSingletonException>(
+      () {
+        ServiceProvider.required<IRequireToBeSingleton>();
+      },
+      "The service IRequireToBeSingleton must be registered with registerSingleton method",
+    );
+
+    _expectException<RegistryMustBeTransientException>(
+      () {
+        ServiceProvider.required<IRequireToBeTransient>();
+      },
+      "The service IRequireToBeTransient must be registered with registerTransient method",
+    );
+
+    ServiceProvider.unregisterAll();
   });
 
   test("Unallow duplicate registration", () {
     _expectException<ServiceAlreadyRegisteredException>(
       () {
-        ServiceProvider.registerTransient<_TestClass>((serviceProvider, platformInfo) => _TestClass());
-        ServiceProvider.registerTransient<_TestClass>((serviceProvider, platformInfo) => _TestClass());
+        ServiceProvider.registerTransient<_TestClass>((optional, required, platform) => _TestClass());
+        ServiceProvider.registerTransient<_TestClass>((optional, required, platform) => _TestClass());
       },
-      "The service _TestClass was already registered as a transactional service",
+      "The service _TestClass was already registered as a transient service",
     );
 
     ServiceProvider.unregisterAll();
 
     _expectException<ServiceAlreadyRegisteredException>(
       () {
-        ServiceProvider.registerSingleton<_TestClass>((serviceProvider, platformInfo) => _TestClass());
-        ServiceProvider.registerSingleton<_TestClass>((serviceProvider, platformInfo) => _TestClass());
+        ServiceProvider.registerSingleton<_TestClass>((optional, required, platform) => _TestClass());
+        ServiceProvider.registerSingleton<_TestClass>((optional, required, platform) => _TestClass());
       },
       "The service _TestClass was already registered as a singleton service",
     );
 
     _expectException<ServiceAlreadyRegisteredException>(
       () {
-        ServiceProvider.registerTransient<_TestClass>((serviceProvider, platformInfo) => _TestClass());
+        ServiceProvider.registerTransient<_TestClass>((optional, required, platform) => _TestClass());
       },
       "The service _TestClass was already registered as a singleton service",
     );
   });
 
   test("Ignore duplicate registration", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass());
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass());
 
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>().value, 0);
+    expect(ServiceProvider.required<_TestClass>().value, 0);
 
-    ServiceProvider.registerTransientIfNotRegistered((serviceProvider, platformInfo) => _MockTestClass(value: 2));
+    ServiceProvider.registerTransientIfNotRegistered((optional, required, platform) => _MockTestClass(value: 2));
 
-    expect(ServiceProvider.instance.getRequiredService<_TestClass>().value, 0);
+    expect(ServiceProvider.required<_TestClass>().value, 0);
   });
 
   test("Registration tests", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "1");
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass());
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "1");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass());
 
-    expect(ServiceProvider.instance.isRegistered<_TestClass>(key: "1"), true);
-    expect(ServiceProvider.instance.isRegistered<_TestClass>(), true);
-    expect(ServiceProvider.instance.isRegistered<_TestClass>(key: "2"), false);
+    expect(ServiceProvider.isRegistered<_TestClass>(key: "1"), true);
+    expect(ServiceProvider.isRegistered<_TestClass>(), true);
+    expect(ServiceProvider.isRegistered<_TestClass>(key: "2"), false);
   });
 
   test("Required services validation", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "1");
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass());
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "1");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass());
 
-    final tc1 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "1");
-    final tc = ServiceProvider.instance.getRequiredService<_TestClass>();
+    final tc1 = ServiceProvider.required<_TestClass>(key: "1");
+    final tc = ServiceProvider.required<_TestClass>();
 
     tc1.value = 1;
     tc.value = 2;
@@ -142,13 +171,13 @@ void main() {
     expect(tc1.value, 1);
     expect(tc.value, 2);
 
-    final tc2 = ServiceProvider.instance.getService<_TestClass>(key: "2");
+    final tc2 = ServiceProvider.optional<_TestClass>(key: "2");
 
     expect(tc2, null);
 
     _expectException<ServiceNotRegisteredException>(
       () {
-        final tc3 = ServiceProvider.instance.getRequiredService<_TestClass>(key: "3");
+        final tc3 = ServiceProvider.required<_TestClass>(key: "3");
 
         expect(tc3, null);
       },
@@ -157,22 +186,22 @@ void main() {
   });
 
   test("Services unregistrations", () {
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass());
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass(), key: "1");
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass());
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass(), key: "1");
 
-    expect(ServiceProvider.instance.getService<_TestClass>()!.value, 0);
-    expect(ServiceProvider.instance.getService<_TestClass>(key: "1")!.value, 0);
+    expect(ServiceProvider.optional<_TestClass>()!.value, 0);
+    expect(ServiceProvider.optional<_TestClass>(key: "1")!.value, 0);
 
     ServiceProvider.unregister<_TestClass>();
 
-    expect(ServiceProvider.instance.getService<_TestClass>(), null);
-    expect(ServiceProvider.instance.getService<_TestClass>(key: "1")!.value, 0);
+    expect(ServiceProvider.optional<_TestClass>(), null);
+    expect(ServiceProvider.optional<_TestClass>(key: "1")!.value, 0);
 
-    ServiceProvider.registerTransient((serviceProvider, platformInfo) => _TestClass());
+    ServiceProvider.registerTransient((optional, required, platform) => _TestClass());
     ServiceProvider.unregister<_TestClass>(key: "1");
 
-    expect(ServiceProvider.instance.getService<_TestClass>()!.value, 0);
-    expect(ServiceProvider.instance.getService<_TestClass>(key: "1"), null);
+    expect(ServiceProvider.optional<_TestClass>()!.value, 0);
+    expect(ServiceProvider.optional<_TestClass>(key: "1"), null);
 
     ServiceProvider.unregisterAll();
 
@@ -185,26 +214,26 @@ void main() {
   });
 
   test("Service without generic argument", () {
-    ServiceProvider.registerSingleton((serviceProvider, platformInfo) => _TestClass());
-    ServiceProvider.registerSingleton((serviceProvider, platformInfo) => _TestClass(), key: "1");
+    ServiceProvider.registerSingleton((optional, required, platform) => _TestClass());
+    ServiceProvider.registerSingleton((optional, required, platform) => _TestClass(), key: "1");
 
     _expectException<ServiceInvalidInferenceException>(
       () {
-        ServiceProvider.registerSingleton((serviceProvider, platformInfo) => null);
+        ServiceProvider.registerSingleton((optional, required, platform) => null);
       },
       "The service with key null was used without specifying the TService generic argument, so I don't know what type to return",
     );
 
     _expectException<ServiceInvalidInferenceException>(
       () {
-        ServiceProvider.instance.getService();
+        ServiceProvider.optional();
       },
       "The service with key null was used without specifying the TService generic argument, so I don't know what type to return",
     );
 
     _expectException<ServiceInvalidInferenceException>(
       () {
-        ServiceProvider.instance.getService(key: "1");
+        ServiceProvider.optional(key: "1");
       },
       "The service with key 1 was used without specifying the TService generic argument, so I don't know what type to return",
     );